requirements.txt
Flask==2.1.0
flask-wtf==0.14.3
flask-jwt-extended==4.3.1
flask-sqlalchemy==2.5.1
email-validator==1.1.3
Werkzeug~=2.0.0
Flask-Mail==0.9.1
flask-cors==3.0.8
flask-bcrypt==0.7.1
Pillow==7.1.2
Flask-Excel==0.0.7
pyexcel-xlsx==0.5.8
pdfkit==0.6.1
requests==2.27.1
